@@ -1,13 +1,12 @@
 {
   "name": "trevordmiller-website",
   "scripts": {
-    "dev": "next",
-    "test": "jest",
-    "lint": "eslint pages components",
     "build": "next build",
-    "verify": "yarn test && yarn lint && yarn build",
+    "dev": "next",
     "start": "next start",
-    "deploy": "git checkout master && git pull && now",
+    "lint": "eslint pages components",
+    "verify": "yarn lint && yarn build",
+    "release": "git checkout master && git pull && now",
     "alias": "now alias $(pbpaste) trevordmiller.com"
   },
   "eslintConfig": {
@@ -43,20 +42,19 @@
     }
   },
   "devDependencies": {
-    "babel-eslint": "^7.1.1",
-    "babel-jest": "^17.0.2",
-    "eslint": "^3.10.2",
+    "babel-eslint": "^7.2.3",
+    "eslint": "^3.19.0",
     "eslint-plugin-react": "^6.10.3",
-    "jest": "^17.0.3",
-    "now": "^4.0.1"
+    "now": "^4.11.2"
   },
   "dependencies": {
     "lodash": "^4.17.4",
-    "next": "^2.1.1",
+    "next": "^2.2.0",
     "nova-colors": "^2.1.5",
-    "prop-types": "^15.5.7",
+    "prop-types": "^15.5.8",
     "react": "^15.5.4",
     "react-dom": "^15.5.4",
-    "react-tabs": "^0.8.2"
+    "react-lazy-load": "^3.0.12",
+    "react-tabs": "^0.8.3"
   }
 }
