@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import Link from 'next/link'
-import {colorValues} from '../utils/styleGuide'
+import {colorValues} from '../../utils/styleGuide'
 
 const Anchor = ({
   href,
