@@ -6,7 +6,7 @@ import {
   fontWeights,
   spacing,
   lineHeights,
-} from '../utils/styleGuide'
+} from '../../utils/styleGuide'
 
 export const levelSpecificStyles = {
   1: {
