@@ -1,6 +1,6 @@
 import React from 'react'
-import {uiGroups, fontSizes, spacing} from '../../../../../utils/styleGuide'
-import {eggheadUrl, githubUrl, twitterUrl, spotifyUrl} from '../../../../../utils/urls'
+import {uiGroups, fontSizes, spacing} from '../../../../../../utils/styleGuide'
+import {eggheadUrl, githubUrl, twitterUrl, spotifyUrl} from '../../../../../../utils/urls'
 import Anchor from '../../../../../Anchor'
 import Icon from '../../../../../Icon'
 
