@@ -1,5 +1,5 @@
 import React from 'react'
-import {spacing, screenSizes} from '../../../utils/styleGuide'
+import {spacing, screenSizes} from '../../../../utils/styleGuide'
 import Info from '../../../Info'
 import Subscribe from './components/Subscribe'
 import Follow from './components/Follow'
@@ -15,7 +15,7 @@ export default () => (
   }}>
     <Info>
       <div>
-        For occasional updates, subscribe to my newsletter or other accounts - no spam
+        Subscribe to my newsletter to be notified when I publish new things. No spam.
       </div>
       <Subscribe />
       <div style={{
