@@ -11,6 +11,9 @@ const key = 'color-scheme-building'
 
 export default () => (
   <Screen
+    baseRouteTitle='Blog'
+    mainVisual={posts[key].mainVisual}
+    pathname={`https://trevordmiller.com/blog/${key}`}
     title={posts[key].title}
     description={posts[key].description}
     showDescription
@@ -54,7 +57,7 @@ export default () => (
               We wanted the project to be easy to maintain, so we decided to create a source module called <Anchor href='https://github.com/trevordmiller/nova-colors'>nova-colors</Anchor> that could contain all of the color values (organized by color meaning), and then use this module in the repo for each editor and tool so that they could all be updated from the same source. We did this for editors (Vim, Atom), terminal emulators (Hyper, iTerm), and other tools (Git, tmux). Here is an example of <Anchor href='https://github.com/trevordmiller/nova-vim/blob/master/src/index.js'>the Nova Vim plugin</Anchor>; you can see how the <Code>src/</Code> directory contains the vim-specific wire-up while using the <Code>nova-colors</Code> package for color values, and then this gets compiled.
             </Paragraph>
             <Paragraph>
-              The nice thing about having <Code>nova-colors</Code> as its own package is it can be used anywhere; for example, <Anchor href='https://github.com/trevordmiller/trevordmiller-website/blob/master/components/utils/styleGuide/index.js#L1-L10'>I'm using it on my own website as part of the style guide module</Anchor>.
+              The nice thing about having <Code>nova-colors</Code> as its own package is it can be used anywhere; for example, <Anchor href='https://github.com/trevordmiller/trevordmiller-website/blob/master/utils/styleGuide/index.js#L1-L10'>I'm using it on my own website as part of the style guide module</Anchor>.
             </Paragraph>
           </div>
         ),
