@@ -1,10 +1,13 @@
 import React, {Component} from 'react'
 import {startCase, first} from 'lodash'
 import Link from 'next/link'
-import {uiGroups, spacing, borderSizes} from '../../../components/utils/styleGuide'
+import {uiGroups, spacing, borderSizes} from '../../../utils/styleGuide'
 import Screen from '../../../components/Screen'
 import Heading from '../../../components/Heading'
 import Button from '../../../components/Button'
+import {projects} from '../'
+
+const key = 'guitar-lessons'
 
 const octave = [1, 0, 2, 0, 3, 4, 0, 5, 0, 6, 0, 7]
 const doubleOctave = octave.concat(octave)
@@ -283,6 +286,9 @@ export default class extends React.Component {
 
     return (
       <Screen
+        baseRouteTitle='Projects'
+        mainVisual={projects[key].mainVisual}
+        pathname={`https://trevordmiller.com/projects/${key}`}
         title='Guitar lessons'
         description='Diatonic (in-key) chord and scale diagrams to help you go beyond tabs and memorizing songs'
         sections={[
